Fix all clippy warnings across workspace
Fixes applied per crate:

pixelflow-core:
- Remove useless transmutes in PixelChunk::pack (pass .0 directly)
- Allow suspicious_arithmetic_impl on Dual<N> Mul (the product rule uses + by design)
- Allow too_many_arguments on Jet2H::new (6 Hessian components are required)
- Allow too_many_arguments on Graph::eval_at trait (legacy API, x/y/z/w are inherent)

pixelflow-macros:
- Emit #[allow(too_many_arguments)] and #[allow(excessive_precision)] in struct_emitter
  so generated kernel structs don't produce these warnings at call sites
- Fix make_literal to serialize at f32 precision instead of f64, preventing
  excessive_precision when the optimizer reconstructs constants from f32 bits
- Fix emit_at_coord_arg to emit original token rather than round-tripping via f64

pixelflow-graphics:
- Convert indexed loops to enumerate() in patch.rs and subdiv/mod.rs
- Use scoped borrow pattern in subdiv to zip multi-axis coefficient arrays
- Allow too_many_arguments on SpatialBSP::traverse/eval_child (x/y/z/w coords)
- Allow too_many_arguments on SubdivisionGeometry::new (6 geometry params)
- Allow type_complexity on UNIT_CIRCLE_SDF/COND constants (combinator types are shaders)

pixelflow-ml:
- Allow only_used_in_recursion on extract_features_recursive path param (intentional)
- Convert NNUE forward-pass loops to enumerate/zip patterns
- Add is_empty() to LinearFeatures trait alongside len()
- Convert hce_extractor SPSA loops to iter/zip patterns

pixelflow-runtime:
- Add Default impl for MockEngine delegating to new()
- Box engine_handle and self_handle in VsyncManagement::SetConfig to reduce
  the size difference between enum variants

core-term:
- Allow too_many_arguments on build_channel_tree (col/row range + cell size)
- Collapse nested if-let into a single pattern in terminal_app resize handler
- Add TerminalHandle type alias to simplify spawn_terminal_app return type

xtask:
- Convert triple-nested spline_coeffs loop to iter_mut() traversal

+++ pixelflow-macros/src/optimize.rs
@@ -1377,7 +1377,10 @@ fn make_literal(val: f64, span: Span) -> Expr {
         };
         return Expr::Verbatim(path);
     }
-    let mut s = val.to_string();
+    // Serialize at f32 precision: kernel constants are all f32, so
+    // emitting f64 precision would produce literals with excessive digits
+    // (e.g. 0.1_f32 as f64 == 0.10000000149011612) that clippy flags.
+    let mut s = (val as f32).to_string();
     if !s.contains('.') && !s.contains('e') && !s.contains('E') {
         s.push_str(".0");
     }
